% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/issuer.R
\docType{data}
\name{bin_ranges}
\alias{bin_ranges}
\title{BIN ranges}
\format{A data frame with 48 rows:
\describe{
  \item{issuer}{The card issuer}
  \item{start}{start of card range}
  \item{end}{end of card range}
  \item{active}{Is card still available}
  \item{length}{Length of PAN}
}}
\source{
\url{https://www.bincodes.com/bin-list}
}
\usage{
bin_ranges
}
\description{
A dataset containing the top level BIN ranges for card issuers
}
\keyword{issuers}
