% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{checkLuhn-package}
\alias{checkLuhn-package}
\title{Functions and data for working with credit and debit cards}
\description{
Contains function for confirm if card numbers are valid and getting
information about the card scheme and issuer
see the repo at https://github.com/adamjdeacon/checkLuhn.
}
\author{
Adam Deacon (\email{adam@deaconsworld.org.uk})
}
