% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkLuhn.R
\name{checkLuhn}
\alias{checkLuhn}
\title{Checks number is Luhn compliant}
\usage{
checkLuhn(number)
}
\arguments{
\item{number}{The number to be checked}
}
\description{
Checks number is Luhn compliant
}
\examples{
checkLuhn('4111 1111 1111 1111')
}
