% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/issuer.R
\name{issuer}
\alias{issuer}
\title{issuer}
\usage{
issuer(number)
}
\arguments{
\item{number}{PAN to look up}
}
\value{
A dataframe of the issuer and if the scheme is active
}
\description{
Attempts to return the issuer and if the scheme is still active.
}
\examples{
issuer('4111 1111 1111 1111')
}
